split connection pool into module
